requirements.txt
py-cord
python-dotenv
cassiopeia
mysql-connector-python